+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Now includes items/stacks worth less than 1c in net worth (previously these were excluded)
 - Added inventory to net worth calculations (with setting in Settings -> Net worth)
 - Added equipment to net worth calculations (with setting in Settings -> Net worth)
+- Added ability to select item-value treshold (with setting Settings -> Net worth)
 - Now shows a separate disconnect-page if the cause is external (e.g if Path of Exile is down)
 - Fixed a bug where a default-setting was not properly set for some options
 - Fixed a bug where analytics screen-views were improperly triggered